+++ CHANGELOG.md
@@ -19,4 +19,5 @@
 
 # 1.2.9 (03/04/2021)
 
--   Do not reset values on useForm state change, because this can cause confusion.
+-   Do not reset values on `useForm` state change, because this can cause confusion.
+-   `form.handleSubmit` helper function.